guidance-for-identifying-diagnosis-codes-from-clinical-notes-on-aws/requirements.txt
aws-cdk-lib==2.141.0
constructs>=10.0.0,<11.0.0
cdklabs.generative-ai-cdk-constructs
cdk-nag